fixed build order
I remember that this project was building correctly when I was actively developing it. But, after leaving it alone for a while, I was told it was not being built correctly and I verified that indeed that was the case. The problem is that some executables that depend on messages were being built before the messages. I integrated [AlvaroFerGar](https://github.com/AlvaroFerGar)'s code from [this issue](https://github.com/GuiRitter/OpenBase/issues/5#issuecomment-411800793) and it seems that this bug has been fixed.